Release lsp-bridge version 2.0.
lsp-bridge 2.0 upgrade note:

Version 2.0 no longer supports corfu and company-mode, because the goal of lsp-bridge is to provide a smooth code completion experience, which relies on the unique push message mechanism of lsp-bridge, and the built-in capf architecture of Emacs cannot satisfy lsp-bridge push mechanism.

Therefore, starting from version 2.0, we created lsp-bridge-fw.el and lsp-bridge-fw-history.el to fork from corfu.el and corfu-history.el respectively, hoping to build on the original code of corfu Develop a new completion framework adapted to the lsp-bridge push mechanism to achieve higher completion performance and completely solve many bugs in the code completion insertion of various languages ​​in existing versions.

The development of lsp-bridge.el has always focused on compatibility with the existing Emacs ecosystem, but in the past two weeks, a large number of bugs related to completion and insertion have been fixed, and it is difficult to completely fix it. The reason is that the capf mechanism cannot satisfy lsp-bridge. The needs of .el: When generating new code completion, lsp-bridg parses the LSP message to generate the latest completion list, corfu has its own candidate word insertion/replacement logic, capf framework also has its own insertion/replacement logic, and then superimposed The three candidate word rendering priorities of LSP (textEdit, insertText and label), especially when the textEdit logic (usually providing the code Snippet function), the insertion/replacement logic of corfu/capf/textEdit are intertwined and conflicted, and eventually lead to The bug cannot be completely fixed when candidate words are inserted.

Please follow the README instructions to configure the latest configuration items, thank you for your support!