Fix issue #831, add heartbeat mechanism for remote usage
If client does not send messages to the server for more than a specified
time(e.g. 2h12min for me), the server will close the socket while the
client can still send messages without error. However, the server will
not receive and handle any more messages.
In order to solve this issue, we add a mechanism to send heartbeats to
the server to keep the connection alive.